refactor: move parse typed data to typed data file
