+++ crates/torii/libp2p/src/server/mod.rs
@@ -7,8 +7,6 @@ use std::time::Duration;
 use std::{fs, io};
 
 use chrono::Utc;
-use crypto_bigint::U256;
-use dojo_types::primitive::Primitive;
 use dojo_types::schema::Ty;
 use dojo_world::contracts::naming::compute_selector_from_names;
 use futures::StreamExt;
@@ -37,7 +35,7 @@ mod events;
 use dojo_world::contracts::model::ModelReader;
 
 use crate::server::events::ServerEvent;
-use crate::typed_data::PrimitiveType;
+use crate::typed_data::{parse_value_to_ty, PrimitiveType};
 use crate::types::Message;
 
 pub(crate) const LOG_TARGET: &str = "torii::relay::server";
@@ -429,206 +427,6 @@ fn ty_keys(ty: &Ty) -> Result<Vec<Felt>, Error> {
     } else {
         Err(Error::InvalidMessageError("Entity is not a struct".to_string()))
     }
-}
-
-pub fn parse_value_to_ty(value: &PrimitiveType, ty: &mut Ty) -> Result<(), Error> {
-    match value {
-        PrimitiveType::Object(object) => match ty {
-            Ty::Struct(struct_) => {
-                for (key, value) in object {
-                    let member =
-                        struct_.children.iter_mut().find(|member| member.name == *key).ok_or_else(
-                            || Error::InvalidMessageError(format!("Member {} not found", key)),
-                        )?;
-
-                    parse_value_to_ty(value, &mut member.ty)?;
-                }
-            }
-            // U256 is an object with two u128 fields
-            // low and high
-            Ty::Primitive(Primitive::U256(u256)) => {
-                let mut low = Ty::Primitive(Primitive::U128(None));
-                let mut high = Ty::Primitive(Primitive::U128(None));
-
-                // parse the low and high fields
-                parse_value_to_ty(&object["low"], &mut low)?;
-                parse_value_to_ty(&object["high"], &mut high)?;
-
-                let low = low.as_primitive().unwrap().as_u128().unwrap();
-                let high = high.as_primitive().unwrap().as_u128().unwrap();
-
-                let mut bytes = [0u8; 32];
-                bytes[..16].copy_from_slice(&high.to_be_bytes());
-                bytes[16..].copy_from_slice(&low.to_be_bytes());
-
-                *u256 = Some(U256::from_be_slice(&bytes));
-            }
-            // an enum is a SNIP-12 compliant object with a single key
-            // where the K is the variant name
-            // and the value is the variant value
-            Ty::Enum(enum_) => {
-                let (option_name, value) = object.first().ok_or_else(|| {
-                    Error::InvalidMessageError("Enum variant not found".to_string())
-                })?;
-
-                enum_.options.iter_mut().for_each(|option| {
-                    if option.name == *option_name {
-                        parse_value_to_ty(value, &mut option.ty).unwrap();
-                    }
-                });
-
-                enum_.set_option(option_name).map_err(|e| {
-                    Error::InvalidMessageError(format!("Failed to set enum option: {}", e))
-                })?;
-            }
-            _ => {
-                return Err(Error::InvalidMessageError(format!(
-                    "Invalid object type for {}",
-                    ty.name()
-                )));
-            }
-        },
-        PrimitiveType::Array(values) => match ty {
-            Ty::Array(array) => {
-                let inner_type = array[0].clone();
-
-                // clear the array, which contains the inner type
-                array.clear();
-
-                // parse each value to the inner type
-                for value in values {
-                    let mut ty = inner_type.clone();
-                    parse_value_to_ty(value, &mut ty)?;
-                    array.push(ty);
-                }
-            }
-            Ty::Tuple(tuple) => {
-                // our array values need to match the length of the tuple
-                if tuple.len() != values.len() {
-                    return Err(Error::InvalidMessageError("Tuple length mismatch".to_string()));
-                }
-
-                for (i, value) in tuple.iter_mut().enumerate() {
-                    parse_value_to_ty(&values[i], value)?;
-                }
-            }
-            _ => {
-                return Err(Error::InvalidMessageError(format!(
-                    "Invalid array type for {}",
-                    ty.name()
-                )));
-            }
-        },
-        PrimitiveType::Number(number) => match ty {
-            Ty::Primitive(primitive) => match *primitive {
-                Primitive::I8(ref mut i8) => {
-                    *i8 = Some(number.as_i64().unwrap() as i8);
-                }
-                Primitive::I16(ref mut i16) => {
-                    *i16 = Some(number.as_i64().unwrap() as i16);
-                }
-                Primitive::I32(ref mut i32) => {
-                    *i32 = Some(number.as_i64().unwrap() as i32);
-                }
-                Primitive::I64(ref mut i64) => {
-                    *i64 = Some(number.as_i64().unwrap());
-                }
-                Primitive::U8(ref mut u8) => {
-                    *u8 = Some(number.as_u64().unwrap() as u8);
-                }
-                Primitive::U16(ref mut u16) => {
-                    *u16 = Some(number.as_u64().unwrap() as u16);
-                }
-                Primitive::U32(ref mut u32) => {
-                    *u32 = Some(number.as_u64().unwrap() as u32);
-                }
-                Primitive::U64(ref mut u64) => {
-                    *u64 = Some(number.as_u64().unwrap());
-                }
-                Primitive::USize(ref mut usize) => {
-                    *usize = Some(number.as_u64().unwrap() as u32);
-                }
-                _ => {
-                    return Err(Error::InvalidMessageError(format!(
-                        "Invalid number type for {}",
-                        ty.name()
-                    )));
-                }
-            },
-            _ => {
-                return Err(Error::InvalidMessageError(format!(
-                    "Invalid number type for {}",
-                    ty.name()
-                )));
-            }
-        },
-        PrimitiveType::Bool(boolean) => {
-            *ty = Ty::Primitive(Primitive::Bool(Some(*boolean)));
-        }
-        PrimitiveType::String(string) => match ty {
-            Ty::Primitive(primitive) => match primitive {
-                Primitive::I8(v) => {
-                    *v = Some(i8::from_str(string).unwrap());
-                }
-                Primitive::I16(v) => {
-                    *v = Some(i16::from_str(string).unwrap());
-                }
-                Primitive::I32(v) => {
-                    *v = Some(i32::from_str(string).unwrap());
-                }
-                Primitive::I64(v) => {
-                    *v = Some(i64::from_str(string).unwrap());
-                }
-                Primitive::I128(v) => {
-                    *v = Some(i128::from_str(string).unwrap());
-                }
-                Primitive::U8(v) => {
-                    *v = Some(u8::from_str(string).unwrap());
-                }
-                Primitive::U16(v) => {
-                    *v = Some(u16::from_str(string).unwrap());
-                }
-                Primitive::U32(v) => {
-                    *v = Some(u32::from_str(string).unwrap());
-                }
-                Primitive::U64(v) => {
-                    *v = Some(u64::from_str(string).unwrap());
-                }
-                Primitive::U128(v) => {
-                    *v = Some(u128::from_str(string).unwrap());
-                }
-                Primitive::USize(v) => {
-                    *v = Some(u32::from_str(string).unwrap());
-                }
-                Primitive::Felt252(v) => {
-                    *v = Some(Felt::from_str(string).unwrap());
-                }
-                Primitive::ClassHash(v) => {
-                    *v = Some(Felt::from_str(string).unwrap());
-                }
-                Primitive::ContractAddress(v) => {
-                    *v = Some(Felt::from_str(string).unwrap());
-                }
-                Primitive::Bool(v) => {
-                    *v = Some(bool::from_str(string).unwrap());
-                }
-                _ => {
-                    return Err(Error::InvalidMessageError("Invalid primitive type".to_string()));
-                }
-            },
-            Ty::ByteArray(s) => {
-                s.clone_from(string);
-            }
-            _ => {
-                return Err(Error::InvalidMessageError(format!(
-                    "Invalid string type for {}",
-                    ty.name()
-                )));
-            }
-        },
-    }
-
-    Ok(())
 }
 
 // Validates the message model
