+++ crates/torii/libp2p/src/typed_data.rs
@@ -1,7 +1,7 @@
 use std::str::FromStr;
 
 use cainome::cairo_serde::ByteArray;
-use crypto_bigint::Encoding;
+use crypto_bigint::{Encoding, U256};
 use dojo_types::primitive::Primitive;
 use dojo_types::schema::Ty;
 use indexmap::IndexMap;
@@ -46,191 +46,6 @@ pub enum PrimitiveType {
     // For JSON numbers. Formed into a Felt
     Number(Number),
 }
-
-pub fn map_ty_to_primitive(ty: &Ty) -> Result<PrimitiveType, Error> {
-    match ty {
-        Ty::Struct(struct_) => {
-            let mut object = IndexMap::new();
-            for member in &struct_.children {
-                object.insert(member.name.clone(), map_ty_to_primitive(&member.ty)?);
-            }
-            Ok(PrimitiveType::Object(object))
-        }
-        Ty::Enum(enum_) => {
-            let mut object = IndexMap::new();
-            let option = enum_.option.map_or(Err(Error::InvalidMessageError("Enum option not found".to_string())), Ok)?;
-            let option = enum_.options.get(option as usize).map_or(Err(Error::InvalidMessageError("Enum option not found".to_string())), Ok)?;
-            object.insert(option.name.clone(), map_ty_to_primitive(&option.ty)?);
-            Ok(PrimitiveType::Object(object))
-        }
-        Ty::Array(array) => {
-            let values: Result<Vec<PrimitiveType>, Error> = array.iter().map(map_ty_to_primitive).collect();
-            Ok(PrimitiveType::Array(values?))
-        }
-        Ty::Tuple(tuple) => {
-            let values: Result<Vec<PrimitiveType>, Error> = tuple.iter().map(map_ty_to_primitive).collect();
-            Ok(PrimitiveType::Array(values?))
-        }
-        Ty::Primitive(primitive) => match primitive {
-            Primitive::Bool(b) => Ok(PrimitiveType::Bool(b.unwrap_or(false))),
-            Primitive::I8(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
-            Primitive::I16(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
-            Primitive::I32(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
-            Primitive::I64(n) => Ok(PrimitiveType::String((n.unwrap_or(0)).to_string())),
-            Primitive::I128(n) => Ok(PrimitiveType::String((n.unwrap_or(0)).to_string())),
-            Primitive::U8(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
-            Primitive::U16(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
-            Primitive::U32(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
-            Primitive::U64(n) => Ok(PrimitiveType::String((n.unwrap_or(0)).to_string())),
-            Primitive::U128(n) => Ok(PrimitiveType::String((n.unwrap_or(0)).to_string())),
-            Primitive::USize(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
-            Primitive::Felt252(f) => Ok(PrimitiveType::String((f.unwrap_or(Felt::ZERO)).to_string())),
-            Primitive::ClassHash(c) => Ok(PrimitiveType::String((c.unwrap_or(Felt::ZERO)).to_string())),
-            Primitive::ContractAddress(c) => Ok(PrimitiveType::String((c.unwrap_or(Felt::ZERO)).to_string())),
-            Primitive::U256(u256) => {
-                let mut object = IndexMap::new();
-                let bytes = u256.map_or([0u8; 32], |u256| u256.to_be_bytes());
-                let high = u128::from_be_bytes(bytes[..16].try_into().unwrap());
-                let low = u128::from_be_bytes(bytes[16..].try_into().unwrap());
-                object.insert("high".to_string(), PrimitiveType::String(high.to_string()));
-                object.insert("low".to_string(), PrimitiveType::String(low.to_string()));
-                Ok(PrimitiveType::Object(object))
-            }
-        },
-        Ty::ByteArray(s) => Ok(PrimitiveType::String(s.clone())),
-    }
-}
-
-fn map_ty_type(types: &mut IndexMap<String, Vec<Field>>, name: &str, ty: Ty) -> Field {
-    match ty {
-        Ty::Primitive(primitive) => match primitive {
-            // map all signed integers to i128
-            Primitive::I8(_)
-            | Primitive::I16(_)
-            | Primitive::I32(_)
-            | Primitive::I64(_)
-            | Primitive::I128(_) => {
-                Field::SimpleType(SimpleField {
-                        name: name.to_string(),
-                        r#type: "i128".to_string(),
-                    })
-            }
-            Primitive::U8(_)
-            | Primitive::U16(_)
-            | Primitive::U32(_)
-            | Primitive::U64(_)
-            | Primitive::USize(_)
-            | Primitive::U128(_) => {
-                Field::SimpleType(SimpleField {
-                        name: name.to_string(),
-                        r#type: "u128".to_string(),
-                    })
-            }
-            Primitive::U256(_) => {
-                Field::SimpleType(SimpleField {
-                        name: name.to_string(),
-                        r#type: "u256".to_string(),
-                    })
-            }
-            Primitive::Bool(_) => {
-                Field::SimpleType(SimpleField {
-                        name: name.to_string(),
-                        r#type: "bool".to_string(),
-                    })
-            }
-            Primitive::Felt252(_) => {
-                Field::SimpleType(SimpleField {
-                        name: name.to_string(),
-                        r#type: "felt".to_string(),
-                    })
-            }
-            Primitive::ClassHash(_) => {
-                Field::SimpleType(SimpleField {
-                        name: name.to_string(),
-                        r#type: "ClassHash".to_string(),
-                    })
-            }
-            Primitive::ContractAddress(_) => {
-                Field::SimpleType(SimpleField {
-                        name: name.to_string(),
-                        r#type: "ContractAddress".to_string(),
-                    })
-            }
-        }
-        Ty::Array(array) => {
-            // if array is empty, we fallback to felt
-            let array_type = if let Some(inner) = array.get(0) {
-                map_ty_type(types, "inner", inner.clone())
-            } else {
-                return Field::SimpleType(SimpleField {
-                    name: name.to_string(),
-                    r#type: "felt".to_string(),
-                });
-            };
-
-            Field::SimpleType(SimpleField {
-                name: name.to_string(),
-                r#type: format!("{}*", match array_type {
-                    Field::SimpleType(simple_field) => simple_field.r#type,
-                    Field::ParentType(parent_field) => parent_field.r#type,
-                }),
-            })
-        }
-        Ty::Struct(struct_ty) => {
-            let mut fields = Vec::new();
-            for member in struct_ty.children.iter() {
-                let field = map_ty_type(types, &member.name, member.ty.clone());
-                fields.push(field);
-            }
-
-            types.insert(struct_ty.name.clone(), fields);
-
-            Field::SimpleType(SimpleField {
-                name: name.to_string(),
-                r#type: struct_ty.name,
-            })
-        }
-        Ty::Enum(enum_ty) => {
-            let mut fields = Vec::new();
-            for option in enum_ty.options.iter() {
-                let field = map_ty_type(types, &option.name, option.ty.clone());
-                fields.push(field);
-            }
-
-            types.insert(enum_ty.name.clone(), fields);
-
-            Field::ParentType(ParentField {
-                name: name.to_string(),
-                r#type: "enum".to_string(),
-                contains: enum_ty.name,
-            })
-        }
-        Ty::Tuple(tuple_ty) => {
-            Field::SimpleType(SimpleField {
-                name: name.to_string(),
-                r#type: format!(
-                    "({})",
-                    tuple_ty
-                        .iter()
-                        .map(|ty| match map_ty_type(types, "inner", ty.clone()) {
-                            Field::SimpleType(simple_field) => simple_field.r#type,
-                            Field::ParentType(parent_field) => parent_field.r#type,
-                        })
-                        .collect::<Vec<String>>()
-                        .join(",")
-                ),
-            })
-        }
-        Ty::ByteArray(_) => {
-            Field::SimpleType(SimpleField {
-                name: name.to_string(),
-                r#type: "string".to_string(),
-            })
-        }
-    }
-}
-
-
 
 fn get_preset_types() -> IndexMap<String, Vec<Field>> {
     let mut types = IndexMap::new();
@@ -337,7 +152,11 @@ pub fn encode_type(name: &str, types: &IndexMap<String, Vec<Field>>) -> Result<S
                             &simple_field.r#type[1..simple_field.r#type.len() - 1]
                                 .split(',')
                                 .map(|t| {
-                                    if !t.is_empty() { format!("\"{}\"", t) } else { t.to_string() }
+                                    if !t.is_empty() {
+                                        format!("\"{}\"", t)
+                                    } else {
+                                        t.to_string()
+                                    }
                                 })
                                 .collect::<Vec<String>>()
                                 .join(",");
@@ -599,6 +418,383 @@ impl Domain {
     }
 }
 
+pub fn parse_value_to_ty(value: &PrimitiveType, ty: &mut Ty) -> Result<(), Error> {
+    match value {
+        PrimitiveType::Object(object) => match ty {
+            Ty::Struct(struct_) => {
+                for (key, value) in object {
+                    let member =
+                        struct_.children.iter_mut().find(|member| member.name == *key).ok_or_else(
+                            || Error::InvalidMessageError(format!("Member {} not found", key)),
+                        )?;
+
+                    parse_value_to_ty(value, &mut member.ty)?;
+                }
+            }
+            // U256 is an object with two u128 fields
+            // low and high
+            Ty::Primitive(Primitive::U256(u256)) => {
+                let mut low = Ty::Primitive(Primitive::U128(None));
+                let mut high = Ty::Primitive(Primitive::U128(None));
+
+                // parse the low and high fields
+                parse_value_to_ty(&object["low"], &mut low)?;
+                parse_value_to_ty(&object["high"], &mut high)?;
+
+                let low = low.as_primitive().unwrap().as_u128().unwrap();
+                let high = high.as_primitive().unwrap().as_u128().unwrap();
+
+                let mut bytes = [0u8; 32];
+                bytes[..16].copy_from_slice(&high.to_be_bytes());
+                bytes[16..].copy_from_slice(&low.to_be_bytes());
+
+                *u256 = Some(U256::from_be_slice(&bytes));
+            }
+            // an enum is a SNIP-12 compliant object with a single key
+            // where the K is the variant name
+            // and the value is the variant value
+            Ty::Enum(enum_) => {
+                let (option_name, value) = object.first().ok_or_else(|| {
+                    Error::InvalidMessageError("Enum variant not found".to_string())
+                })?;
+
+                enum_.options.iter_mut().for_each(|option| {
+                    if option.name == *option_name {
+                        parse_value_to_ty(value, &mut option.ty).unwrap();
+                    }
+                });
+
+                enum_.set_option(option_name).map_err(|e| {
+                    Error::InvalidMessageError(format!("Failed to set enum option: {}", e))
+                })?;
+            }
+            _ => {
+                return Err(Error::InvalidMessageError(format!(
+                    "Invalid object type for {}",
+                    ty.name()
+                )));
+            }
+        },
+        PrimitiveType::Array(values) => match ty {
+            Ty::Array(array) => {
+                let inner_type = array[0].clone();
+
+                // clear the array, which contains the inner type
+                array.clear();
+
+                // parse each value to the inner type
+                for value in values {
+                    let mut ty = inner_type.clone();
+                    parse_value_to_ty(value, &mut ty)?;
+                    array.push(ty);
+                }
+            }
+            Ty::Tuple(tuple) => {
+                // our array values need to match the length of the tuple
+                if tuple.len() != values.len() {
+                    return Err(Error::InvalidMessageError("Tuple length mismatch".to_string()));
+                }
+
+                for (i, value) in tuple.iter_mut().enumerate() {
+                    parse_value_to_ty(&values[i], value)?;
+                }
+            }
+            _ => {
+                return Err(Error::InvalidMessageError(format!(
+                    "Invalid array type for {}",
+                    ty.name()
+                )));
+            }
+        },
+        PrimitiveType::Number(number) => match ty {
+            Ty::Primitive(primitive) => match *primitive {
+                Primitive::I8(ref mut i8) => {
+                    *i8 = Some(number.as_i64().unwrap() as i8);
+                }
+                Primitive::I16(ref mut i16) => {
+                    *i16 = Some(number.as_i64().unwrap() as i16);
+                }
+                Primitive::I32(ref mut i32) => {
+                    *i32 = Some(number.as_i64().unwrap() as i32);
+                }
+                Primitive::I64(ref mut i64) => {
+                    *i64 = Some(number.as_i64().unwrap());
+                }
+                Primitive::U8(ref mut u8) => {
+                    *u8 = Some(number.as_u64().unwrap() as u8);
+                }
+                Primitive::U16(ref mut u16) => {
+                    *u16 = Some(number.as_u64().unwrap() as u16);
+                }
+                Primitive::U32(ref mut u32) => {
+                    *u32 = Some(number.as_u64().unwrap() as u32);
+                }
+                Primitive::U64(ref mut u64) => {
+                    *u64 = Some(number.as_u64().unwrap());
+                }
+                Primitive::USize(ref mut usize) => {
+                    *usize = Some(number.as_u64().unwrap() as u32);
+                }
+                _ => {
+                    return Err(Error::InvalidMessageError(format!(
+                        "Invalid number type for {}",
+                        ty.name()
+                    )));
+                }
+            },
+            _ => {
+                return Err(Error::InvalidMessageError(format!(
+                    "Invalid number type for {}",
+                    ty.name()
+                )));
+            }
+        },
+        PrimitiveType::Bool(boolean) => {
+            *ty = Ty::Primitive(Primitive::Bool(Some(*boolean)));
+        }
+        PrimitiveType::String(string) => match ty {
+            Ty::Primitive(primitive) => match primitive {
+                Primitive::I8(v) => {
+                    *v = Some(i8::from_str(string).unwrap());
+                }
+                Primitive::I16(v) => {
+                    *v = Some(i16::from_str(string).unwrap());
+                }
+                Primitive::I32(v) => {
+                    *v = Some(i32::from_str(string).unwrap());
+                }
+                Primitive::I64(v) => {
+                    *v = Some(i64::from_str(string).unwrap());
+                }
+                Primitive::I128(v) => {
+                    *v = Some(i128::from_str(string).unwrap());
+                }
+                Primitive::U8(v) => {
+                    *v = Some(u8::from_str(string).unwrap());
+                }
+                Primitive::U16(v) => {
+                    *v = Some(u16::from_str(string).unwrap());
+                }
+                Primitive::U32(v) => {
+                    *v = Some(u32::from_str(string).unwrap());
+                }
+                Primitive::U64(v) => {
+                    *v = Some(u64::from_str(string).unwrap());
+                }
+                Primitive::U128(v) => {
+                    *v = Some(u128::from_str(string).unwrap());
+                }
+                Primitive::USize(v) => {
+                    *v = Some(u32::from_str(string).unwrap());
+                }
+                Primitive::Felt252(v) => {
+                    *v = Some(Felt::from_str(string).unwrap());
+                }
+                Primitive::ClassHash(v) => {
+                    *v = Some(Felt::from_str(string).unwrap());
+                }
+                Primitive::ContractAddress(v) => {
+                    *v = Some(Felt::from_str(string).unwrap());
+                }
+                Primitive::Bool(v) => {
+                    *v = Some(bool::from_str(string).unwrap());
+                }
+                _ => {
+                    return Err(Error::InvalidMessageError("Invalid primitive type".to_string()));
+                }
+            },
+            Ty::ByteArray(s) => {
+                s.clone_from(string);
+            }
+            _ => {
+                return Err(Error::InvalidMessageError(format!(
+                    "Invalid string type for {}",
+                    ty.name()
+                )));
+            }
+        },
+    }
+
+    Ok(())
+}
+
+pub fn map_ty_to_primitive(ty: &Ty) -> Result<PrimitiveType, Error> {
+    match ty {
+        Ty::Struct(struct_) => {
+            let mut object = IndexMap::new();
+            for member in &struct_.children {
+                object.insert(member.name.clone(), map_ty_to_primitive(&member.ty)?);
+            }
+            Ok(PrimitiveType::Object(object))
+        }
+        Ty::Enum(enum_) => {
+            let mut object = IndexMap::new();
+            let option = enum_
+                .option
+                .map_or(Err(Error::InvalidMessageError("Enum option not found".to_string())), Ok)?;
+            let option = enum_
+                .options
+                .get(option as usize)
+                .map_or(Err(Error::InvalidMessageError("Enum option not found".to_string())), Ok)?;
+            object.insert(option.name.clone(), map_ty_to_primitive(&option.ty)?);
+            Ok(PrimitiveType::Object(object))
+        }
+        Ty::Array(array) => {
+            let values: Result<Vec<PrimitiveType>, Error> =
+                array.iter().map(map_ty_to_primitive).collect();
+            Ok(PrimitiveType::Array(values?))
+        }
+        Ty::Tuple(tuple) => {
+            let values: Result<Vec<PrimitiveType>, Error> =
+                tuple.iter().map(map_ty_to_primitive).collect();
+            Ok(PrimitiveType::Array(values?))
+        }
+        Ty::Primitive(primitive) => match primitive {
+            Primitive::Bool(b) => Ok(PrimitiveType::Bool(b.unwrap_or(false))),
+            Primitive::I8(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
+            Primitive::I16(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
+            Primitive::I32(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
+            Primitive::I64(n) => Ok(PrimitiveType::String((n.unwrap_or(0)).to_string())),
+            Primitive::I128(n) => Ok(PrimitiveType::String((n.unwrap_or(0)).to_string())),
+            Primitive::U8(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
+            Primitive::U16(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
+            Primitive::U32(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
+            Primitive::U64(n) => Ok(PrimitiveType::String((n.unwrap_or(0)).to_string())),
+            Primitive::U128(n) => Ok(PrimitiveType::String((n.unwrap_or(0)).to_string())),
+            Primitive::USize(n) => Ok(PrimitiveType::Number(Number::from(n.unwrap_or(0)))),
+            Primitive::Felt252(f) => {
+                Ok(PrimitiveType::String((f.unwrap_or(Felt::ZERO)).to_string()))
+            }
+            Primitive::ClassHash(c) => {
+                Ok(PrimitiveType::String((c.unwrap_or(Felt::ZERO)).to_string()))
+            }
+            Primitive::ContractAddress(c) => {
+                Ok(PrimitiveType::String((c.unwrap_or(Felt::ZERO)).to_string()))
+            }
+            Primitive::U256(u256) => {
+                let mut object = IndexMap::new();
+                let bytes = u256.map_or([0u8; 32], |u256| u256.to_be_bytes());
+                let high = u128::from_be_bytes(bytes[..16].try_into().unwrap());
+                let low = u128::from_be_bytes(bytes[16..].try_into().unwrap());
+                object.insert("high".to_string(), PrimitiveType::String(high.to_string()));
+                object.insert("low".to_string(), PrimitiveType::String(low.to_string()));
+                Ok(PrimitiveType::Object(object))
+            }
+        },
+        Ty::ByteArray(s) => Ok(PrimitiveType::String(s.clone())),
+    }
+}
+
+fn map_ty_type(types: &mut IndexMap<String, Vec<Field>>, name: &str, ty: Ty) -> Field {
+    match ty {
+        Ty::Primitive(primitive) => match primitive {
+            // map all signed integers to i128
+            Primitive::I8(_)
+            | Primitive::I16(_)
+            | Primitive::I32(_)
+            | Primitive::I64(_)
+            | Primitive::I128(_) => Field::SimpleType(SimpleField {
+                name: name.to_string(),
+                r#type: "i128".to_string(),
+            }),
+            Primitive::U8(_)
+            | Primitive::U16(_)
+            | Primitive::U32(_)
+            | Primitive::U64(_)
+            | Primitive::USize(_)
+            | Primitive::U128(_) => Field::SimpleType(SimpleField {
+                name: name.to_string(),
+                r#type: "u128".to_string(),
+            }),
+            Primitive::U256(_) => Field::SimpleType(SimpleField {
+                name: name.to_string(),
+                r#type: "u256".to_string(),
+            }),
+            Primitive::Bool(_) => Field::SimpleType(SimpleField {
+                name: name.to_string(),
+                r#type: "bool".to_string(),
+            }),
+            Primitive::Felt252(_) => Field::SimpleType(SimpleField {
+                name: name.to_string(),
+                r#type: "felt".to_string(),
+            }),
+            Primitive::ClassHash(_) => Field::SimpleType(SimpleField {
+                name: name.to_string(),
+                r#type: "ClassHash".to_string(),
+            }),
+            Primitive::ContractAddress(_) => Field::SimpleType(SimpleField {
+                name: name.to_string(),
+                r#type: "ContractAddress".to_string(),
+            }),
+        },
+        Ty::Array(array) => {
+            // if array is empty, we fallback to felt
+            let array_type = if let Some(inner) = array.get(0) {
+                map_ty_type(types, "inner", inner.clone())
+            } else {
+                return Field::SimpleType(SimpleField {
+                    name: name.to_string(),
+                    r#type: "felt".to_string(),
+                });
+            };
+
+            Field::SimpleType(SimpleField {
+                name: name.to_string(),
+                r#type: format!(
+                    "{}*",
+                    match array_type {
+                        Field::SimpleType(simple_field) => simple_field.r#type,
+                        Field::ParentType(parent_field) => parent_field.r#type,
+                    }
+                ),
+            })
+        }
+        Ty::Struct(struct_ty) => {
+            let mut fields = Vec::new();
+            for member in struct_ty.children.iter() {
+                let field = map_ty_type(types, &member.name, member.ty.clone());
+                fields.push(field);
+            }
+
+            types.insert(struct_ty.name.clone(), fields);
+
+            Field::SimpleType(SimpleField { name: name.to_string(), r#type: struct_ty.name })
+        }
+        Ty::Enum(enum_ty) => {
+            let mut fields = Vec::new();
+            for option in enum_ty.options.iter() {
+                let field = map_ty_type(types, &option.name, option.ty.clone());
+                fields.push(field);
+            }
+
+            types.insert(enum_ty.name.clone(), fields);
+
+            Field::ParentType(ParentField {
+                name: name.to_string(),
+                r#type: "enum".to_string(),
+                contains: enum_ty.name,
+            })
+        }
+        Ty::Tuple(tuple_ty) => Field::SimpleType(SimpleField {
+            name: name.to_string(),
+            r#type: format!(
+                "({})",
+                tuple_ty
+                    .iter()
+                    .map(|ty| match map_ty_type(types, "inner", ty.clone()) {
+                        Field::SimpleType(simple_field) => simple_field.r#type,
+                        Field::ParentType(parent_field) => parent_field.r#type,
+                    })
+                    .collect::<Vec<String>>()
+                    .join(",")
+            ),
+        }),
+        Ty::ByteArray(_) => {
+            Field::SimpleType(SimpleField { name: name.to_string(), r#type: "string".to_string() })
+        }
+    }
+}
+
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct TypedData {
     pub types: IndexMap<String, Vec<Field>>,
@@ -618,17 +814,13 @@ impl TypedData {
         Self { types, primary_type: primary_type.to_string(), domain, message }
     }
 
-    pub fn from_model(
-        model: Ty,
-        domain: Domain,
-    ) -> Result<Self, Error> {
+    pub fn from_model(model: Ty, domain: Domain) -> Result<Self, Error> {
         let model = model.as_struct().expect("Model must be a struct");
         let mut types = IndexMap::new();
         let mut values = IndexMap::new();
         values.insert("model".to_string(), PrimitiveType::String(model.name.clone()));
-        
-        let mut model_values = IndexMap::new();
 
+        let mut model_values = IndexMap::new();
 
         let mut fields = Vec::new();
         for member in model.children.iter() {
